# backend/static/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Masterblog API",
    "description": "API for a simple blog application",
    "version": "1.0.0"
  },
  "paths": {
    "/api/posts": {
      "get": {
        "summary": "Get all posts",
        "parameters": [
          {
            "name": "sort",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Field to sort by"
          },
          {
            "name": "direction",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Sort direction"
          }
        ],
        "responses": {
          "200": {
            "description": "A list of posts"
          }
        }
      },
      "post": {
        "summary": "Create a new post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "content": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Post created successfully"
          }
        }
      }
    },
    "/api/posts/{id}": {
      "put": {
        "summary": "Update a post",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "content": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Post updated successfully"
          }
        }
      },
      "delete": {
        "summary": "Delete a post",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Post deleted successfully"
          }
        }
      }
    },
    "/api/posts/search": {
      "get": {
        "summary": "Search for posts",
        "parameters": [
          {
            "name": "title",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "content",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A list of posts that match the search criteria"
          }
        }
      }
    }
  }
}